allow execution against a hidden notebook editor.
